pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
tensorflow = "^2.8.0"
tensorflow-hub = "^0.12.0"
keras = "^2.8.0"
Pillow = "^9.5.0"
Flask = "^2.3.2"
Flask-Cors = "^3.0.10"
aiosqlite = "0.19.0"
"discord.py-self" = "2.0.0"
aiohttp = "3.8.4"
pyarmor = "8.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"